updated middleware

+++ api/tasks.js
@@ -39,31 +39,36 @@ router.route("/:id").get(async (req, res) => {
     res.send(task)
 })
 
-router.route("/:id").delete(async (req, res) => {
+router.route("/:id").delete(verifyToken, async (req, res) => {
     const id = Number(req.params.id)
     if (!isValidId(id)) {
-    return res.status(400).send({ error: "ID must be a positive integer" });
+        return res.status(400).send({ error: "ID must be a positive integer" });
   }
-    const task = await deleteTask(id)
+    const task = await getTask(id)
     if(!task){
-        res.status(404).send({error: "Task not found"})
+        return res.status(404).send({error: "Task not found"})
     }
-    res.sendStatus(204)
+    if (req.user.id !== task.user_id) {
+        return res.status(403).send({error: "Forbidden: not authorized to delete this task."});
+    }
+    await deleteTask(id)
+    res.sendStatus(204).send("Task Deleted")
 })
 
-router.route("/:id").put(async (req, res) => {
+router.route("/:id").put(verifyToken, async (req, res) => {
     const id = Number(req.params.id)
     if(!req.body){
         return res.status(400).send({error: "Missing body"})
     }
-    const {title, done} = req.body
 
+    const {title, done} = req.body
     if(!title || done == null ){
         return res.status(400).send({error: "Missing rquired fields"})
     }
     if (!isValidId(id)) {
-    return res.status(400).send({ error: "ID must be a positive integer" });
+        return res.status(400).send({ error: "ID must be a positive integer" });
   }
+
     const task = await getTask(id)
     if(!task){
         return res.status(404).send({error: "Task not found"})
